+++ docs/release-notes.md
@@ -6,6 +6,7 @@
 - fix health tread 
 - change initial status for document
 - added RemoveFolderInputModel and ClearOutDocumentInputModel models
+- added AvailabilityML for ML services
 
 # 0.0.82